Time: 13 ms (8.68%) | Memory: 21 MB (17.35%) - LeetSync

+++ 2148-minimum-number-of-moves-to-seat-everyone/minimum-number-of-moves-to-seat-everyone.cpp
@@ -3,7 +3,7 @@ public:
     int minMovesToSeat(vector<int>& seats, vector<int>& students) {
         sort(seats.begin(),seats.end());
         sort(students.begin(),students.end());
-        long long cont=0;
+        int cont=0;
         for(int i=0;i<seats.size();i++)
         {
             cont+=abs(students[i]-seats[i]);
